refactor(cache): move array helpers to src/core/apolloClient/cacheHelpers
